+++ eng/smoke-test-package.ps1
@@ -1,6 +1,6 @@
 param(
     [string]$Configuration = "Release",
-    [string]$Version = "0.9.0-alpha.13",
+    [string]$Version = "0.9.0-alpha.14",
     [switch]$UseMajorRollForward
 )